+++ src/components/root/PageRoot.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode } from 'react';
 import { ThemeProvider } from 'styled-components';
 import { Normalize } from 'styled-normalize';
 import { GlobalStyle } from './GlobalStyle';
-import { theme } from '../../data/theme';
+import { theme } from '../../theme/theme';
 import { Header } from '../common/header/Header';
 import { Footer } from '../common/footer/Footer';
