+++ src/theme/color.ts
@@ -1,0 +1,11 @@
+export const createColor = () => ({
+    blue100: '#ebf8ff',
+    blue400: '#63b3ed',
+    blue500: '#4299e1',
+    blue700: '#2b6cb0',
+    blue900: '#2a4365',
+    gray400: '#cbd5e0',
+    gray500: '#a0aec0',
+    gray700: '#4a5568',
+    white: '#fff',
+});
