+++ src/theme/shadow.ts
@@ -1,0 +1,3 @@
+export const createShadow = () => ({
+    md: '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
+});
