+++ src/theme/theme.ts
@@ -1,0 +1,9 @@
+import { createColor } from './color';
+import { createShadow } from './shadow';
+import { createFontSize } from './fontSize';
+
+export const theme = {
+    color: createColor(),
+    fontSize: createFontSize(),
+    shadow: createShadow(),
+};
